requirements.txt
h5py==3.14.0
joblib==1.5.2
lightgbm==4.6.0
matplotlib==3.10.6
medpy==0.5.2
mlflow==3.5.0
mlflow_skinny==3.5.0
mlflow_tracing==3.5.0
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.3
optuna==4.5.0
pandas==2.3.2
plotly==6.3.1
pypalettes==0.2.1
python-dotenv==1.1.1
pyyaml==6.0.2
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
scikit-image==0.25.2
statsmodels==0.14.5
surface_distance==0.1
tensorflow==2.20.0
tqdm==4.67.1
